+++ example/lib/screens/forms.dart
@@ -130,20 +130,24 @@ class _FormsState extends State<Forms> {
           ),
         ]),
         const SizedBox(height: 20),
-        TextBox(
-          maxLines: null,
-          controller: _clearController,
-          suffixMode: OverlayVisibilityMode.always,
-          minHeight: 100,
-          suffix: _clearController.text.isEmpty
-              ? null
-              : IconButton(
-                  icon: const Icon(FluentIcons.chrome_close),
-                  onPressed: () {
-                    _clearController.clear();
-                  },
-                ),
-          placeholder: 'Text box with clear button',
+        ConstrainedBox(
+          constraints: const BoxConstraints(maxHeight: 200),
+          child: TextFormBox(
+            maxLines: null,
+            controller: _clearController,
+            suffixMode: OverlayVisibilityMode.always,
+            minHeight: 100,
+            expands: true,
+            suffix: _clearController.text.isEmpty
+                ? null
+                : IconButton(
+                    icon: const Icon(FluentIcons.chrome_close),
+                    onPressed: () {
+                      _clearController.clear();
+                    },
+                  ),
+            placeholder: 'Text box with clear button',
+          ),
         ),
         const SizedBox(height: 20),
         TextBox(