+++ lib/src/controls/navigation/navigation_view/pane.dart
@@ -85,6 +85,7 @@ class NavigationPane with Diagnosticable {
     this.customPane,
     this.menuButton,
     this.scrollController,
+    this.leading,
     this.indicator = const StickyNavigationIndicator(),
   }) : assert(selected == null || selected >= 0);
 
@@ -167,6 +168,9 @@ class NavigationPane with Diagnosticable {
   /// the scrolling and keep the state of the scroll when the
   /// display mode is toggled.
   final ScrollController? scrollController;
+
+  /// The leading Widget for the Pane
+  final Widget? leading;
 
   /// A function called when building the navigation indicator
   final Widget? indicator;
@@ -322,12 +326,19 @@ class NavigationPaneSize {
   /// maxWidth must be greater or equal than minWidth.
   final double? openMaxWidth;
 
+  /// The height of the header in NavigationPane.
+  ///
+  /// Only used when NavigationPane mode is open.
+  /// If the value is null, [_kOneLineTileHeight] is used.
+  final double? headerHeight;
+
   const NavigationPaneSize({
     this.topHeight,
     this.compactWidth,
     this.openWidth,
     this.openMinWidth,
     this.openMaxWidth,
+    this.headerHeight,
   }) : assert(
           openMinWidth == null ||
               openMaxWidth == null ||
@@ -376,13 +387,11 @@ class _TopNavigationPane extends StatefulWidget {
   _TopNavigationPane({
     required this.pane,
     this.listKey,
-    this.scrollbarKey,
     this.appBar,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
   final GlobalKey? listKey;
-  final GlobalKey? scrollbarKey;
   final NavigationAppBar? appBar;
 
   @override
@@ -432,8 +441,14 @@ class _TopNavigationPaneState extends State<_TopNavigationPane> {
       child: Row(children: [
         Expanded(
           child: Row(children: [
-            if (widget.appBar != null)
-              NavigationAppBar.buildLeading(context, widget.appBar!),
+            if (widget.pane.leading != null)
+              Padding(
+                padding: const EdgeInsets.symmetric(
+                  horizontal: 8.0,
+                  vertical: 6.0,
+                ),
+                child: widget.pane.leading!,
+              ),
             if (widget.pane.header != null)
               Padding(
                 padding: const EdgeInsets.symmetric(
@@ -517,14 +532,12 @@ class _CompactNavigationPane extends StatelessWidget {
     required this.pane,
     this.paneKey,
     this.listKey,
-    this.scrollbarKey,
     this.onToggle,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
   final Key? paneKey;
   final GlobalKey? listKey;
-  final GlobalKey? scrollbarKey;
   final VoidCallback? onToggle;
 
   Widget _buildItem(BuildContext context, NavigationPaneItem item) {
@@ -596,17 +609,12 @@ class _CompactNavigationPane extends StatelessWidget {
               ),
             ),
           Expanded(
-            child: Scrollbar(
-              key: scrollbarKey,
-              controller: pane.scrollController,
-              isAlwaysShown: false,
-              child: ListView(
-                key: listKey,
-                primary: true,
-                children: pane.items.map((item) {
-                  return _buildItem(context, item);
-                }).toList(),
-              ),
+            child: ListView(
+              key: listKey,
+              primary: true,
+              children: pane.items.map((item) {
+                return _buildItem(context, item);
+              }).toList(),
             ),
           ),
           ListView(
@@ -629,7 +637,6 @@ class _OpenNavigationPane extends StatefulWidget {
     required this.theme,
     this.paneKey,
     this.listKey,
-    this.scrollbarKey,
     this.onToggle,
     this.onItemSelected,
   }) : super(key: pane.key);
@@ -637,7 +644,6 @@ class _OpenNavigationPane extends StatefulWidget {
   final NavigationPane pane;
   final Key? paneKey;
   final GlobalKey? listKey;
-  final GlobalKey? scrollbarKey;
   final VoidCallback? onToggle;
   final VoidCallback? onItemSelected;
 
@@ -741,7 +747,7 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
             margin: widget.pane.autoSuggestBox != null
                 ? EdgeInsets.zero
                 : topPadding,
-            height: kOneLineTileHeight,
+            height: widget.pane.size?.headerHeight ?? kOneLineTileHeight,
             child: () {
               if (widget.pane.header != null) {
                 return Row(children: [
@@ -767,22 +773,17 @@ class _OpenNavigationPaneState extends State<_OpenNavigationPane>
               child: widget.pane.autoSuggestBox!,
             ),
           Expanded(
-            child: Scrollbar(
-              key: widget.scrollbarKey,
-              controller: widget.pane.scrollController,
-              isAlwaysShown: false,
-              child: ListView(
-                key: widget.listKey,
-                primary: true,
-                children: widget.pane.items.map((item) {
-                  return _OpenNavigationPane.buildItem(
-                    context,
-                    widget.pane,
-                    item,
-                    widget.onItemSelected,
-                  );
-                }).toList(),
-              ),
+            child: ListView(
+              key: widget.listKey,
+              primary: true,
+              children: widget.pane.items.map((item) {
+                return _OpenNavigationPane.buildItem(
+                  context,
+                  widget.pane,
+                  item,
+                  widget.onItemSelected,
+                );
+              }).toList(),
             ),
           ),
           ListView(