+++ lib/src/controls/surfaces/flyout/flyout.dart
@@ -1,9 +1,29 @@
+import 'dart:async';
+
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 
 import '../../../utils/popup.dart';
 
 export 'controller.dart';
+
+part 'content.dart';
+part 'menu.dart';
+
+const kDefaultLongHoverDuration = Duration(milliseconds: 400);
+
+/// Where the flyout will be placed vertically relativelly the child
+enum FlyoutPosition {
+  /// The flyout will be above the child, if there is enough space available
+  above,
+
+  /// The flyout will be below the child, if there is enough space available
+  below,
+
+  /// The flyout will be by the side of the child, if there is enough space
+  /// available
+  side,
+}
 
 /// How the flyout will be placed relatively to the child
 enum FlyoutPlacement {
@@ -35,6 +55,9 @@ enum FlyoutOpenMode {
 
   /// The flyout will opened when the user hover the child
   hover,
+
+  /// The flyout will be opened when the user long hover the child
+  longHover,
 
   /// The flyout will opened when the user press the child
   press,
@@ -68,20 +91,27 @@ class Flyout extends StatefulWidget {
     this.horizontalOffset = 10.0,
     this.placement = FlyoutPlacement.center,
     this.openMode = FlyoutOpenMode.none,
+    this.position = FlyoutPosition.above,
+    this.longHoverDuration = kDefaultLongHoverDuration,
+    this.onOpen,
+    this.onClose,
   }) : super(key: key);
 
   /// The child that will be attached to the flyout.
-  ///
-  /// Usually a [FlyoutContent]
   final Widget child;
 
-  /// The content that will be displayed on the route
+  /// The content that will be displayed on the flyout.
+  ///
+  /// Usually a [FlyoutContent] is used
   final WidgetBuilder content;
 
   /// Holds the state of the flyout. Can be useful to open or close the flyout
   /// programatically.
   ///
-  /// Call `controller.dispose` to clean up resources when no longer necessary
+  /// Call `controller.dispose()` to clean up resources when no longer necessary
+  ///
+  /// See also:
+  ///   * [openMode], which can open the flyout on hover, press and long press
   final FlyoutController? controller;
 
   /// The vertical gap between the [child] and the displayed flyout.
@@ -90,9 +120,9 @@ class Flyout extends StatefulWidget {
   /// The horizontal gap between the [child] and the displayed flyout.
   final double horizontalOffset;
 
-  /// How the flyout will be placed relatively to the [child].
+  /// How the flyout will be placed horizontally relatively to the [child].
   ///
-  /// Defaults to center
+  /// Defaults to [FlyoutPlacement.center]
   final FlyoutPlacement placement;
 
   /// How the flyout will be opened by the end-user without needing to use a
@@ -100,6 +130,22 @@ class Flyout extends StatefulWidget {
   ///
   /// Defaults to none
   final FlyoutOpenMode openMode;
+
+  /// The duration of the hover if [openMode] is [FlyoutOpenMode.longHover].
+  ///
+  /// 800 milliseconds are used by default
+  final Duration longHoverDuration;
+
+  /// Where the flyout will be placed vertically relatively to the child
+  ///
+  /// Defaults to [FlyoutPosition.above]
+  final FlyoutPosition position;
+
+  /// Called when the flyout is opened, either by [controller] or [openMode]
+  final VoidCallback? onOpen;
+
+  /// Called when the flyout is closed, either by [controller] or by the user
+  final VoidCallback? onClose;
 
   @override
   _FlyoutState createState() => _FlyoutState();
@@ -110,9 +156,14 @@ class Flyout extends StatefulWidget {
     properties
       ..add(DiagnosticsProperty<FlyoutController>('controller', controller))
       ..add(DoubleProperty(
-        'verticalOffset',
+        'vertical offset',
         verticalOffset,
         defaultValue: 24.0,
+      ))
+      ..add(DoubleProperty(
+        'horizontal offset',
+        horizontalOffset,
+        defaultValue: 10.0,
       ))
       ..add(EnumProperty<FlyoutPlacement>(
         'placement',
@@ -123,6 +174,16 @@ class Flyout extends StatefulWidget {
         'open mode',
         openMode,
         defaultValue: FlyoutOpenMode.none,
+      ))
+      ..add(EnumProperty<FlyoutPosition>(
+        'position',
+        position,
+        defaultValue: FlyoutPosition.above,
+      ))
+      ..add(DiagnosticsProperty<Duration>(
+        'long hover duration',
+        longHoverDuration,
+        defaultValue: kDefaultLongHoverDuration,
       ));
   }
 }
@@ -131,22 +192,20 @@ class _FlyoutState extends State<Flyout> {
   final popupKey = GlobalKey<PopUpState>();
 
   late FlyoutController controller;
+  Timer? longHoverTimer;
 
   @override
   void initState() {
     super.initState();
-
     controller = widget.controller ?? FlyoutController();
-
     controller.addListener(_handleStateChanged);
   }
 
   @override
   void didUpdateWidget(covariant Flyout oldWidget) {
     super.didUpdateWidget(oldWidget);
-
     if (oldWidget.controller == null && widget.controller != null) {
-      // Dispose the current controller
+      // Dispose the current controller, which was created locally
       controller.dispose();
 
       // Assign to the new controller
@@ -156,11 +215,16 @@ class _FlyoutState extends State<Flyout> {
   }
 
   void _handleStateChanged() {
+    if (!mounted) return;
     final isOpen = popupKey.currentState?.isOpen ?? false;
     if (!isOpen && controller.isOpen) {
-      popupKey.currentState?.openPopup();
+      popupKey.currentState?.openPopup().then((value) {
+        widget.onClose?.call();
+      });
+      widget.onOpen?.call();
     } else if (isOpen && controller.isClosed) {
       Navigator.pop(context);
+      widget.onClose?.call();
     }
   }
 
@@ -171,6 +235,8 @@ class _FlyoutState extends State<Flyout> {
     if (widget.controller == null) {
       controller.dispose();
     }
+    longHoverTimer?.cancel();
+    longHoverTimer = null;
     super.dispose();
   }
 
@@ -183,6 +249,7 @@ class _FlyoutState extends State<Flyout> {
       verticalOffset: widget.verticalOffset,
       horizontalOffset: widget.horizontalOffset,
       placement: widget.placement,
+      position: widget.position,
     );
 
     switch (widget.openMode) {
@@ -192,6 +259,17 @@ class _FlyoutState extends State<Flyout> {
         return MouseRegion(
           opaque: false,
           onEnter: (event) => controller.open(),
+          child: popup,
+        );
+      case FlyoutOpenMode.longHover:
+        return MouseRegion(
+          opaque: true,
+          onEnter: (event) {
+            longHoverTimer = Timer(widget.longHoverDuration, controller.open);
+          },
+          onExit: (event) {
+            if (longHoverTimer?.isActive ?? false) longHoverTimer?.cancel();
+          },
           child: popup,
         );
       case FlyoutOpenMode.press:
@@ -207,197 +285,5 @@ class _FlyoutState extends State<Flyout> {
           child: popup,
         );
     }
-  }
-}
-
-/// The content of the flyout.
-///
-/// See also:
-///
-///   * [Flyout]
-///   * [FlyoutListTile]
-class FlyoutContent extends StatelessWidget {
-  /// Creates a flyout content
-  const FlyoutContent({
-    Key? key,
-    required this.child,
-    this.color,
-    this.shape,
-    this.padding = const EdgeInsets.all(8.0),
-    this.shadowColor = Colors.black,
-    this.elevation = 8,
-    this.constraints,
-  }) : super(key: key);
-
-  final Widget child;
-
-  /// The background color of the box.
-  final Color? color;
-
-  /// The shape to fill the [color] of the box.
-  final ShapeBorder? shape;
-
-  /// Empty space to inscribe around the [child]
-  final EdgeInsetsGeometry padding;
-
-  /// The shadow color.
-  final Color shadowColor;
-
-  /// The z-coordinate relative to the box at which to place this physical
-  /// object.
-  final double elevation;
-
-  /// Additional constraints to apply to the child.
-
-  final BoxConstraints? constraints;
-
-  @override
-  Widget build(BuildContext context) {
-    assert(debugCheckHasFluentTheme(context));
-    final ThemeData theme = FluentTheme.of(context);
-    return PhysicalModel(
-      elevation: elevation,
-      color: Colors.transparent,
-      shadowColor: shadowColor,
-      child: Container(
-        constraints: constraints,
-        decoration: ShapeDecoration(
-          color: color ?? theme.menuColor,
-          shape: shape ??
-              RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(6.0),
-                side: BorderSide(
-                  width: 0.25,
-                  color: theme.inactiveBackgroundColor,
-                ),
-              ),
-        ),
-        padding: padding,
-        child: DefaultTextStyle(
-          style: theme.typography.body ?? const TextStyle(),
-          child: child,
-        ),
-      ),
-    );
-  }
-}
-
-/// A tile that is used inside of [FlyoutContent]
-class FlyoutListTile extends StatelessWidget {
-  /// Creates a flyout list tile.
-  const FlyoutListTile({
-    Key? key,
-    this.onPressed,
-    this.tooltip,
-    this.icon,
-    required this.text,
-    this.trailing,
-    this.focusNode,
-    this.autofocus = false,
-    this.semanticLabel,
-  }) : super(key: key);
-
-  final VoidCallback? onPressed;
-
-  /// The tile tooltip text
-  final String? tooltip;
-
-  /// The leading widget.
-  ///
-  /// Usually an [Icon]
-  final Widget? icon;
-
-  /// The title widget.
-  ///
-  /// Usually a [Text]
-  final Widget text;
-
-  /// The leading widget.
-  final Widget? trailing;
-
-  /// {@macro flutter.widgets.Focus.focusNode}
-  final FocusNode? focusNode;
-
-  /// {@macro flutter.widgets.Focus.autofocus}
-  final bool autofocus;
-
-  /// {@macro fluent_ui.controls.inputs.HoverButton.semanticLabel}
-  final String? semanticLabel;
-
-  @override
-  Widget build(BuildContext context) {
-    assert(debugCheckHasFluentTheme(context));
-    return HoverButton(
-      key: key,
-      onPressed: onPressed,
-      focusNode: focusNode,
-      autofocus: autofocus,
-      semanticLabel: semanticLabel,
-      builder: (context, states) {
-        final theme = FluentTheme.of(context);
-        final radius = BorderRadius.circular(4.0);
-
-        Widget content = Container(
-          decoration: BoxDecoration(
-            color: ButtonThemeData.uncheckedInputColor(theme, states),
-            borderRadius: radius,
-          ),
-          padding: const EdgeInsetsDirectional.only(
-            top: 4.0,
-            bottom: 4.0,
-            start: 10.0,
-            end: 8.0,
-          ),
-          child: Row(mainAxisSize: MainAxisSize.min, children: [
-            if (icon != null)
-              Padding(
-                padding: const EdgeInsetsDirectional.only(end: 10.0),
-                child: IconTheme.merge(
-                  data: const IconThemeData(size: 16.0),
-                  child: icon!,
-                ),
-              ),
-            Flexible(
-              child: Padding(
-                padding: const EdgeInsetsDirectional.only(end: 10.0),
-                child: DefaultTextStyle(
-                  child: text,
-                  style: TextStyle(
-                    inherit: false,
-                    fontSize: 14.0,
-                    letterSpacing: -0.15,
-                    color: theme.inactiveColor,
-                  ),
-                ),
-              ),
-            ),
-            if (trailing != null)
-              DefaultTextStyle(
-                child: trailing!,
-                style: TextStyle(
-                  inherit: false,
-                  fontSize: 12.0,
-                  color: theme.borderInputColor,
-                  height: 0.7,
-                ),
-              ),
-          ]),
-        );
-
-        if (tooltip != null) {
-          content = Tooltip(message: tooltip, child: content);
-        }
-
-        return Padding(
-          padding: const EdgeInsets.only(bottom: 5.0),
-          child: FocusBorder(
-            focused: states.isFocused,
-            renderOutside: true,
-            style: FocusThemeData(borderRadius: radius),
-            child: content,
-          ),
-        );
-      },
-    );
   }
 }